Split 'em if you got 'em
