+++ languages/python/sqlalchemy-oso/sqlalchemy_oso/oso.py
@@ -1,6 +1,6 @@
 from oso import Oso, OsoError
 from .auth import register_models
-from .roles import PolarRoles
+from .roles import OsoRoles
 
 
 class SQLAlchemyOso(Oso):
@@ -37,10 +37,9 @@ class SQLAlchemyOso(Oso):
         """
         if self._roles_enabled:
             raise OsoError("Roles feature already enabled.")
-        self._roles = PolarRoles(
+        self._roles = OsoRoles(
             oso=self,
             user_model=user_model,
-            sqlalchemy_base=self.base,
             session_maker=session_maker,
         )
         self._roles_enabled = True
